- use graphql handler for exposing /graph endpoint - update glide yaml - refactor graphql code to separate file
