+++ glide.yaml
@@ -66,3 +66,5 @@ import:
   version: ~1.1.1
 - package: github.com/graphql-go/graphql
   version: ^0.7.5
+- package: github.com/graphql-go/handler
+  version: ^0.2.1
